+++ src/Spinorama/index.tsx
@@ -13,47 +13,60 @@ import type { SpinoramaProps, SpinoramaSettings } from './index.types'
 
 const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 	// Props
-	const { duration, className, children } = props
+	const { duration, animateDuration, ease, className, children } = props
 
 	// Variables
 	const container = React.useRef<HTMLDivElement>(null)
+	const [selected, setSelected] = React.useState<number>(0)
+	const [totalItems, setTotalItems] = React.useState<number>(0)
 
 	// Settings
 	const settings: SpinoramaSettings = {
-		duration: duration || 10000
+		duration: duration || 10000,
+		animateDuration: (animateDuration || 1000) / 1000,
+		ease: ease || 'power1.inOut'
 	}
 
 	// Callbacks
-	const { contextSafe } = useGSAP({ scope: container })
+	const { contextSafe } = useGSAP(
+		() => {
+			// Set interval
+			const interval = setInterval(nextItem, settings.duration)
 
-	const nextItem = React.useCallback(() => {
-		contextSafe(() => {
-			if (container.current) {
-				// Get items
-				const items = container.current.querySelectorAll('.spinorama-wrapper')
+			// Find total items
+			if (container.current && !totalItems) {
+				const items = container.current.querySelectorAll('.spinorama-item')
 
-				if (items.length) {
-					// Animate
-					gsap.to(items, {
-						xPercent: -100,
-						duration: 1,
-						ease: 'power2.inOut'
-					})
-				}
+				if (items.length) setTotalItems(items.length)
 			}
-		})
-	}, [])
 
-	React.useEffect(() => {
-		// Call interval
-		const interval = setInterval(nextItem, settings.duration)
+			return () => {
+				// Clear interval
+				clearInterval(interval)
+			}
+		},
+		{ scope: container, dependencies: [container, selected, totalItems] }
+	)
 
-		// Cleanup effect
-		return () => {
-			// Clear the interval when the component is unmounted
-			clearInterval(interval)
+	const nextItem = contextSafe(() => {
+		if (container.current && totalItems) {
+			// Set next Item
+			const nextIndex = (selected + 1) % totalItems
+			setSelected(nextIndex)
+
+			// Find items
+			const items = container.current.querySelectorAll('.spinorama-items')
+
+			// Check items and animate it
+			if (items.length) {
+				gsap.to(items, {
+					xPercent: 100 * nextIndex,
+					duration: settings.animateDuration,
+					ease: settings.ease
+				})
+			}
 		}
-	}, [nextItem, settings.duration])
+	})
 
 	return (
 		<Box ref={container} {...props} className={`spinorama${className ? ` ${className}` : ''}`}>